requirement.txt
torch==1.7.0
torchvision==0.8.1
imageio==2.9.0
scikit-image==0.17.2
pandas==1.1.1
matplotlib==3.3.0
tqdm==4.47.0
Pillow==7.2.0
tensorboardX==2.1
tensorflow-gpu==1.14.0
